+++ app/src/androidTest/java/com/example/topgmeals/IngredientStorageTest.java
@@ -11,9 +11,9 @@ import android.widget.EditText;
 import androidx.test.ext.junit.runners.AndroidJUnit4;
 import androidx.test.rule.ActivityTestRule;
 
-import com.example.topgmeals.mealplan.ingredientstorage.AddEditIngredientActivity;
-import com.example.topgmeals.mealplan.ingredientstorage.Ingredient;
-import com.example.topgmeals.mealplan.ingredientstorage.IngredientStorage;
+import com.example.topgmeals.ingredientstorage.AddEditIngredientActivity;
+import com.example.topgmeals.ingredientstorage.Ingredient;
+import com.example.topgmeals.ingredientstorage.IngredientStorage;
 import com.robotium.solo.Solo;
 
 import org.junit.Before;
