+++ app/src/main/java/com/example/topgmeals/ingredientstorage/AddEditIngredientActivity.java
@@ -1,4 +1,4 @@
-package com.example.topgmeals.mealplan.ingredientstorage;
+package com.example.topgmeals.ingredientstorage;
 
 import android.app.AlertDialog;
 import android.app.DatePickerDialog;
