+++ app/src/main/java/com/example/topgmeals/ingredientstorage/IngredientAdapter.java
@@ -1,4 +1,4 @@
-package com.example.topgmeals.mealplan.ingredientstorage;
+package com.example.topgmeals.ingredientstorage;
 
 import android.view.LayoutInflater;
 import android.view.View;
