+++ app/src/main/java/com/example/topgmeals/mealplan/MealPlan.java
@@ -9,7 +9,7 @@ import android.widget.ExpandableListView;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.topgmeals.R;
-import com.example.topgmeals.mealplan.ingredientstorage.IngredientStorage;
+import com.example.topgmeals.ingredientstorage.IngredientStorage;
 import com.example.topgmeals.recipebook.RecipeBook;
 import com.example.topgmeals.shoppinglist.ShoppingList;
 import com.google.firebase.auth.FirebaseAuth;
