+++ app/src/main/java/com/example/topgmeals/recipebook/RecipeBook.java
@@ -15,7 +15,7 @@ import androidx.annotation.RequiresApi;
 import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.topgmeals.R;
-import com.example.topgmeals.mealplan.ingredientstorage.IngredientStorage;
+import com.example.topgmeals.ingredientstorage.IngredientStorage;
 import com.example.topgmeals.mealplan.MealPlan;
 import com.example.topgmeals.shoppinglist.ShoppingList;
 import com.google.firebase.auth.FirebaseAuth;
