+++ app/src/main/java/com/example/topgmeals/shoppinglist/ShoppingListFinish.java
@@ -21,9 +21,9 @@ import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.topgmeals.R;
-import com.example.topgmeals.mealplan.ingredientstorage.AddEditIngredientActivity;
-import com.example.topgmeals.mealplan.ingredientstorage.Ingredient;
-import com.example.topgmeals.mealplan.ingredientstorage.IngredientAdapter;
+import com.example.topgmeals.ingredientstorage.AddEditIngredientActivity;
+import com.example.topgmeals.ingredientstorage.Ingredient;
+import com.example.topgmeals.ingredientstorage.IngredientAdapter;
 import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.firebase.auth.FirebaseAuth;
