+++ app/src/main/res/layout/activity_recipee_information.xml
@@ -58,7 +58,7 @@
             android:layout_marginTop="10dp"
             android:layout_marginRight="10dp"
             android:layout_weight="3"
-            android:inputType="numberSigned"
+            android:inputType="number"
             android:fontFamily="@font/satoshi_medium"
             android:textColor="@color/black"
             android:textSize="18sp" />
@@ -87,7 +87,7 @@
             android:layout_height="wrap_content"
             android:layout_marginRight="10dp"
             android:layout_weight="3"
-            android:inputType="numberSigned"
+            android:inputType="number"
             android:fontFamily="@font/satoshi_medium"
             android:textColor="@color/black"
             android:textSize="18sp" />
