+++ app/src/main/res/layout/add_edit_ingredient.xml
@@ -112,7 +112,7 @@
             android:layout_width="150dp"
             android:layout_marginRight="10dp"
             android:layout_weight="3"
-            android:inputType="numberSigned"
+            android:inputType="number"
             android:fontFamily="@font/satoshi_medium"
             android:textColor="@color/black"
             android:textSize="18sp" />
