+++ app/src/main/res/layout/add_recipee.xml
@@ -59,7 +59,7 @@
             android:layout_marginTop="10dp"
             android:layout_marginRight="10dp"
             android:layout_weight="3"
-            android:inputType="numberSigned"
+            android:inputType="number"
             android:fontFamily="@font/satoshi_medium"
             android:textColor="@color/black"
             android:textSize="18sp" />
@@ -86,7 +86,7 @@
             android:id="@+id/serving_editText"
             android:layout_width="250dp"
             android:layout_height="wrap_content"
-            android:inputType="numberSigned"
+            android:inputType="number"
             android:layout_marginRight="10dp"
             android:layout_weight="3"
             android:fontFamily="@font/satoshi_medium"
